match/requirements.txt
flask
apscheduler
pyyaml
psycopg2
python-Levenshtein
fuzzywuzzy